Changed BlockElt to unsigned long
Hope is that this will allow creation of more than 2^32 Param's for
the same G. Required small mods to blocks, KLread. Changed Elt in
Poset to unsigned long to allow poset of FPP facets in E8.

+++ sources/Atlas.h
@@ -443,7 +443,7 @@ namespace atlas {
   using blocks::Block_base;
   using blocks::Block;
   using blocks::common_block;
-  typedef unsigned int BlockElt;
+  typedef unsigned long int BlockElt;
   typedef std::vector<BlockElt> BlockEltList;
   typedef std::pair<BlockElt,BlockElt> BlockEltPair;
   typedef std::vector<BlockEltPair> BlockEltPairList;
